Move string_to_money/money_to_string from util to localisation
